Filters: organization, typos, remove old tag handling
